image_story_generator/requirements.txt
Pillow==9.2.0
transformers==4.21.2
fastapi[all]
tensorflow-cpu==2.2.0
numpy==1.23.2
opencv_python_headless==4.5.3.56